Fix a few compiler warnings.

+++ subprocess.c
@@ -49,7 +49,7 @@ FILE *subprocess_get_result(struct subprocess_handle **handle, pid_t pid) {
     int fd = open(name, O_RDONLY);
     char buf[1024];
     size_t n = pread(fd, buf, 1024, 0);
-    fprintf(stderr, "file read:\n%.*s===\n", n, buf);
+    fprintf(stderr, "file read:\n%.*s===\n", (int)n, buf);
     FILE *res = fdopen(fd, "r");
     
     unlink(name);
